// src/shared/api/orval/types/availableScheduleResponse.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Smarter Store API
 * Smarter Store 백엔드 API 문서
 * OpenAPI spec version: 1.0.0
 */
import type { TicketOptionWithRemainingSeatsResponse } from './ticketOptionWithRemainingSeatsResponse';

/**
 * 예매 가능 회차 응답 DTO (잔여석 포함)
 */
export interface AvailableScheduleResponse {
  /** 회차 ID */
  id: number;
  /** 공연 날짜 및 시간 */
  showDateTime: string;
  /** 등급별 잔여석 정보 */
  ticketOptions: TicketOptionWithRemainingSeatsResponse[];
}