// src/shared/api/orval/types/createPerformanceScheduleRequest.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Smarter Store API
 * Smarter Store 백엔드 API 문서
 * OpenAPI spec version: 1.0.0
 */
import type { TicketOptionRequest } from './ticketOptionRequest';

/**
 * 공연 회차 생성 요청 DTO
 */
export interface CreatePerformanceScheduleRequest {
  /** 공연 날짜 및 시간 */
  showDateTime: string;
  /** 티켓 판매 시작 일시 */
  saleStartDateTime: string;
  /** 회차별 티켓 가격 옵션 목록 */
  ticketOptions: TicketOptionRequest[];
}